* move services to packages * renames currency history dao
