+++ src/main/java/com/rybka/dao/HibernateUtil.java
@@ -36,10 +36,4 @@ public class HibernateUtil {
         }
         return sessionFactory;
     }
-
-    public static void shutdown() {
-        if (registry != null) {
-            StandardServiceRegistryBuilder.destroy(registry);
-        }
-    }
 }
