+++ src/main/java/com/rybka/service/connector/BaseCurrencyExchangeConnector.java
@@ -1,4 +1,4 @@
-package com.rybka.service;
+package com.rybka.service.connector;
 
 import com.rybka.model.ExchangeResponse;
