+++ src/main/java/com/rybka/service/connector/PrimeExchangeRateConnector.java
@@ -1,4 +1,4 @@
-package com.rybka.service;
+package com.rybka.service.connector;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.rybka.config.CurrencyAPIConstants;
